Phase vocoder OK code needs ti be clean Pitch shifting is ok @8192 double buffers

+++ src/main/java/fr/gnark/sound/domain/physics/PitchShift.java
@@ -29,11 +29,9 @@ public class PitchShift {
         }
         PolynomialSplineFunction function = linearInterpolator.interpolate(x, reconstructedSignal);
         for (int i = 0; i < result.length; i++) {
-            result[i] = function.value(((int)(i * ratio)));
+            result[i] = function.value(i * ratio);
         }
 
         return result;
     }
-
-
 }
